bluehost_deployment/requirements.txt
# Core Dependencies
discord.py>=2.3.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
aiomysql>=0.1.0

# API and Data Processing
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Image Processing
Pillow>=10.0.0

# Testing and Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
ipython>=8.14.0
jupyter>=1.0.0
notebook>=7.0.0

# Web Framework
flask>=2.3.0

# Machine Learning and Analytics
scikit-learn>=1.3.0
torch>=2.0.0
nltk>=3.8.0
textblob>=0.17.0

# WebSocket Support
websockets>=11.0.0

# GraphQL Support
graphene>=3.2.0

# Advanced Security
cryptography>=41.0.0
bcrypt>=4.0.0
